server/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
PyMySQL==1.1.1
aiomysql==0.2.0
cryptography==44.0.3
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
pydantic==2.11.7
python-multipart==0.0.20
httpx==0.28.1
